requirements.txt
bentoml
joblib
pip-chill
torchvision
tqdm
wincertstore
torchsummary
matplotlib
opencv-python
fastapi
uvicorn